add endpoint and connected to frontend
