+++ backend/api/views/admin.py
@@ -1,37 +1,37 @@
 from flask import Blueprint, request, jsonify
 from firebase_admin import auth as firebase_admin_auth
 from api.core import create_response, serialize_list, logger
-from api.models import MentorProfile, Users, VerifiedEmail, Admin
+from api.models import MentorProfile, MenteeProfile, Users, VerifiedEmail, Admin
 from api.utils.require_auth import admin_only
+from api.utils.request_utils import get_profile_model
+from api.utils.constants import Account
 import csv
 import io
 
 
 admin = Blueprint("admin", __name__)  # initialize blueprint
 
-# DELETE request for specific mentor based on id
-
-
-@admin.route("/mentor/<string:mentor_id>", methods=["DELETE"])
+# DELETE request for specific account based on id
+@admin.route("/<int:account_type>/<string:id>", methods=["DELETE"])
 @admin_only
-def delete_mentor(mentor_id):
+def delete_account(role, id):
     try:
-        mentor = MentorProfile.objects.get(id=mentor_id)
+        account = get_profile_model(role).objects.get(id=id)
     except:
-        msg = "No mentors currently exist with ID " + mentor_id
+        msg = "No account currently exist with ID " + id
         logger.info(msg)
         return create_response(status=422, message=msg)
 
-    user_id = mentor.user_id
-    firebase_uid = mentor.firebase_uid
-    email = mentor.email
+    user_id = account.user_id
+    firebase_uid = account.firebase_uid
+    email = account.email
     login = None
 
     if not firebase_uid:
         try:
             login = Users.objects.get(id=user_id.id)
         except:
-            msg = "No mentors currently exist with user_id " + user_id
+            msg = "No account currently exist with user_id " + user_id
             logger.info(msg)
             return create_response(status=422, message=msg)
 
@@ -40,7 +40,7 @@ def delete_mentor(mentor_id):
             try:
                 verified = VerifiedEmail.objects.get(email=email)
             except:
-                msg = "No verified mentors currently exist with email " + email
+                msg = "No verified account currently exist with email " + email
                 logger.info(msg)
 
         login.delete()
@@ -50,10 +50,10 @@ def delete_mentor(mentor_id):
         login = firebase_admin_auth.get_user(firebase_uid)
 
         if not login.email_verified:
-            msg = "No verified mentors currently exist with email " + email
+            msg = "No verified account currently exist with email " + email
             logger.info(msg)
 
-    mentor.delete()
+    account.delete()
     return create_response(status=200, message="Successful deletion")
