+++ frontend/src/components/AdminDataTable.js
@@ -4,7 +4,7 @@ import { LinkOutlined, DeleteOutlined } from "@ant-design/icons";
 
 import "./css/AdminAccountData.scss";
 import { formatLinkForHref } from "utils/misc";
-import { MENTEE_PROFILE, MENTOR_PROFILE } from "utils/consts";
+import { ACCOUNT_TYPE, MENTEE_PROFILE, MENTOR_PROFILE } from "utils/consts";
 
 const { Column } = Table;
 
@@ -55,7 +55,11 @@ function AdminDataTable({ data, deleteAccount }) {
           <Popconfirm
             title={`Are you sure you want to delete ${data.name}?`}
             onConfirm={() => {
-              deleteAccount(data.id, data.name);
+              deleteAccount(
+                data.id,
+                data.isMentee ? ACCOUNT_TYPE.MENTEE : ACCOUNT_TYPE.MENTOR,
+                data.name
+              );
             }}
             onCancel={() =>
               message.info(`No deletion has been for ${data.name}`)
